Updated anno using np.repeat and formatting
I removed your method of adding annotation and replaced it with using
np.repeat. I use the black code formatter for formatting code and
this is the result of it being run on your code.

I saved the annotations as a different csv using "_anno.csv" appended to the
file name.